+++ History.md
@@ -1,3 +1,11 @@
+...
+==================
+  * Change the default json content type to application/json
+  * add a basic-return event for exchanges to catch simple
+    basicReturn messages.
+  * cleaner and more manageable handling of server channel closes
+    from Glen Mailer.
+
 0.1.0 / 2011-07-30
 ==================